Добавила в __main__ method для post по вакансиям и перенесла get по вакансиям
